requirements.txt
streamlit==1.29.0
torch==2.1.0
transformers==4.36.2
pillow==10.0.0
pdf2image==1.16.3
pytesseract==0.3.10
pdfplumber==0.10.1